+++ src/router/index.ts
@@ -1,12 +1,10 @@
 import { createRouter, createWebHistory } from "vue-router";
+import bindHooks from "./hooks";
 import Signin from "@/pages/signin/Signin.vue";
 import Signup from "@/pages/signup/Signup.vue";
 import Home from "@/pages/home/Home.vue";
 import Error from "@/pages/error/Error.vue";
 import Users from "@/pages/users/Users.vue";
-import { useAccountStore } from "@/stores/useAccountStore";
-import { AccountApi } from "@/api-clients/common";
-import type { RouteLocationNormalized, NavigationGuardNext } from "vue-router";
 /**
  * ルートの meta 情報
  * @typedef {Object} RouteMeta
@@ -53,36 +51,6 @@ const router = createRouter({
   routes,
 });
 
-// 認証チェック
-const checkAuthentication = async (
-  to: RouteLocationNormalized,
-  from: RouteLocationNormalized,
-  next: NavigationGuardNext
-) => {
-  if (to.meta.skipAuth) {
-    next();
-    return;
-  }
-
-  const api = new AccountApi();
-  try {
-    const data = await api.getAccount();
-    // アカウント情報をストアに保存
-    const account = useAccountStore();
-    account.set(data);
-    next();
-  } catch (e) {
-    // セッションエラー時の処理
-    const currentPath = window.location.pathname;
-    const expiresDate = new Date();
-    expiresDate.setTime(expiresDate.getTime() + 3 * 60 * 1000);
-    document.cookie = `previousPath=${currentPath}; expires=${expiresDate.toUTCString()}; path=/`;
-
-    next({ name: "Signin" });
-  }
-};
-
-// ナビゲーションガード
-router.beforeEach(checkAuthentication);
+bindHooks(router);
 
 export default router;
